Validasi Struktur email dan Perubahan models
