+++ app/models/__init__.py
@@ -1,0 +1,7 @@
+from .entry import Entry
+from .user import User
+
+__all__ = [
+    "User",
+    "Entry"
+]
