+++ app/models/entry.py
@@ -1,0 +1,10 @@
+from app import db, app
+
+class Entry(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String(64), index=True, nullable=False)
+    description = db.Column(db.String(120), index=True, nullable=False)
+    status = db.Column(db.Boolean, default=False)
+
+    def __repr__(self):
+        return self.title
