+++ app/templates/login.html
@@ -2,7 +2,7 @@
 
 {% block content%}
     <div>
-        <form action="{{ url_for('login') }}" method="POST">
+        <form action="{{ url_for('login') }}" method="POST" novalidate>
             <div class="form-group">
                 <label for="inputEmail">Email / Name:</label><br>
                 <input id="inputEmail" type="text" name="email" class="form-control w-25" placeholder="Email / Name" aria-describedby="emailHep">
